Used the flex Shorthand Property
Added the CSS property flex to both #box-1 and #box-2. Gave #box-1 the values so its flex-grow is 2, its flex-shrink is 2, and its flex-basis is 150px. Gave #box-2 the values so its flex-grow is 1, its flex-shrink is 1, and its flex-basis is 150px.

These values will cause #box-1 to grow to fill the extra space at twice the rate of #box-2 when the container is greater than 300px and shrink at twice the rate of #box-2 when the container is less than 300px. 300px is the combined size of the flex-basis values of the two boxes.